Fix return type of LCMP/[DF]CMP[LG] instructions.
The original version of the code lumped them with instructions
such as FADD, which take two arguments of the same type (e.g.,
float in the case of FADD) and return a result of the same type.

For LCMP, DCMPL, DCMPG, FCMPL and FCMPG, this is incorrect: While they
take two arguments of identical type, they always return an int.