Fixes Issue 188 - Crashes around MethodHandleImpl (#189)
This commit fixes issue #188 with these changes:

Ensure that, when an unwrapped object is passed to an unistrumented wrapper (to a parameter of type java.lang.Object), that the object is wrapped before passing to the instrumented code. Exclude "Unsafe" from this handling, as it is necessary to still be able to use putReference etc. to set an underlying field
For @CallerSensitive methods, collect the callerClass in the uninstrumented wrapper and pass it into the instrumented code